+++ yandex/data_source_yandex_compute_instance_group.go
@@ -47,13 +47,16 @@ func dataSourceYandexComputeInstanceGroup() *schema.Resource {
 										Type:     schema.TypeFloat,
 										Computed: true,
 									},
-
 									"cores": {
 										Type:     schema.TypeInt,
 										Computed: true,
 									},
 
 									"core_fraction": {
+										Type:     schema.TypeInt,
+										Computed: true,
+									},
+									"gpus": {
 										Type:     schema.TypeInt,
 										Computed: true,
 									},