+++ yandex/resource_yandex_compute_instance_test.go
@@ -54,6 +54,29 @@ func TestAccComputeInstance_basic1(t *testing.T) {
 	})
 }
 
+func TestAccComputeInstance_Gpus(t *testing.T) {
+	var instance compute.Instance
+	var instanceName = fmt.Sprintf("instance-test-gpus-%s", acctest.RandString(10))
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:     func() { testAccPreCheck(t) },
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckComputeInstanceDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccComputeInstance_gpus(instanceName),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckComputeInstanceExists(
+						"yandex_compute_instance.foobar", &instance),
+					testAccCheckComputeInstanceHasInstanceID(&instance, "yandex_compute_instance.foobar"),
+					testAccCheckComputeInstanceHasGpus(&instance, 1),
+				),
+			},
+			computeInstanceImportStep(),
+		},
+	})
+}
+
 func TestAccComputeInstance_basic2(t *testing.T) {
 	t.Parallel()
 
@@ -943,6 +966,16 @@ func testAccCheckComputeInstanceHasResources(instance *compute.Instance, cores, 
 	}
 }
 
+func testAccCheckComputeInstanceHasGpus(instance *compute.Instance, gpus int64) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+		resources := instance.GetResources()
+		if resources.Gpus != gpus {
+			return fmt.Errorf("Wrong instance Gpus resource: expected %d, got %d", gpus, resources.Gpus)
+		}
+		return nil
+	}
+}
+
 func testAccCheckComputeInstanceHasSubnet(instance *compute.Instance) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
 		for _, i := range instance.NetworkInterfaces {
@@ -1063,6 +1096,56 @@ resource "yandex_vpc_network" "inst-test-network" {}
 
 resource "yandex_vpc_subnet" "inst-test-subnet" {
   zone           = "ru-central1-a"
+  network_id     = "${yandex_vpc_network.inst-test-network.id}"
+  v4_cidr_blocks = ["192.168.0.0/24"]
+}
+`, instance)
+}
+
+func testAccComputeInstance_gpus(instance string) string {
+	return fmt.Sprintf(`
+data "yandex_compute_image" "ubuntu" {
+  family = "ubuntu-1804-lts"
+}
+
+resource "yandex_compute_instance" "foobar" {
+  name        = "%s"
+  description = "testAccComputeInstance_basic"
+  zone        = "ru-central1-b"
+  platform_id = "gpu-standard-v1"
+
+  resources {
+    cores  = 8
+    memory = 96
+    gpus   = 1
+  }
+
+  boot_disk {
+    initialize_params {
+      size     = 4
+      image_id = "${data.yandex_compute_image.ubuntu.id}"
+    }
+  }
+
+  network_interface {
+    subnet_id = "${yandex_vpc_subnet.inst-test-subnet.id}"
+  }
+
+  metadata = {
+    foo = "bar"
+    baz = "qux"
+  }
+
+  labels = {
+    my_key       = "my_value"
+    my_other_key = "my_other_value"
+  }
+}
+
+resource "yandex_vpc_network" "inst-test-network" {}
+
+resource "yandex_vpc_subnet" "inst-test-subnet" {
+  zone           = "ru-central1-b"
   network_id     = "${yandex_vpc_network.inst-test-network.id}"
   v4_cidr_blocks = ["192.168.0.0/24"]
 }